fix: defective good combination manually and related reports
